JsonIgnore `changedSchemas` & `changedElements`.
The reason that we can safely omit these fields from the result JSON is because their data is referenced in `changedOperations`, so it serves no benefit in retaining the redundant data.

`changedElements` is especially problematic, as it's an interface method implemented by ~25 OpenAPI objects, which simply stores a list of each & every changed data that can grow exponentially when dealing with nested structures. We don't need this, as this data exists in `changedOperations` in a more meaningful structure.